Split out the target rewrite pass into its own source file. Add the ability to test if a subprogram is recursive or not.
