+++ flang/lib/Optimizer/CMakeLists.txt
@@ -16,6 +16,7 @@ add_flang_library(FIROptimizer
   CodeGen/CodeGen.cpp
   CodeGen/PreCGRewrite.cpp
   CodeGen/Target.cpp
+  CodeGen/TargetRewrite.cpp
 
   Transforms/ControlFlowConverter.cpp
   Transforms/CSE.cpp
